refactor: 에뮬레이터 컴포넌트 분리
