+++ emulator/src/routes/emulator/CarLogList.tsx
@@ -1,0 +1,13 @@
+function CarLogList() {
+    return (
+        <div>
+            <ul>
+                <li>Car Log 1</li>
+                <li>Car Log 2</li>
+                <li>Car Log 3</li>
+            </ul>
+        </div>
+    )
+}
+
+export default CarLogList;
