+++ emulator/src/routes/emulator/Sidebar.tsx
@@ -1,0 +1,8 @@
+function SideBar() {
+    return (
+        <div className="min-w-40">
+        </div>
+    )
+}
+
+export default SideBar;
